requirements.txt
certifi==2019.3.9
Django==2.2
pytz==2019.1
sqlparse==0.3.0
djangorestframework==3.9.2
djangorestframework_simplejwt==4.3.0
markdown==3.1
django-filter==2.1.0
django-modeladmin-reorder==0.3.1
raven==6.10.0
django-jet==1.0.8
twilio==6.26.2
coreapi==2.3.3
PyYAML==5.1
pyotp==2.2.7
gunicorn==19.9.0
whitenoise==4.1.2
django-redis==4.10.0
hiredis==1.0.0
psycopg2==2.8.1
psycopg2cffi==2.8.1
django-phonenumber-field==2.4.0
django-enumfields==1.0.0
phonenumbers==8.10.10
django-randomslugfield==0.3.0
Celery==4.3.0
watchdog==0.9.0
django-easy-audit==1.1
django-annoying==0.10.4
Pillow==6.0.0
django-imagekit==4.0.2
boto3==1.9.174
django-storages==1.7.1
django-celery-results==1.1.2
sorl-thumbnail==12.5.0
sorl-thumbnail-serializer-field==0.2.1
nexmo==2.4.0
simplejson==3.16.0
django-crequest==2018.5.11
onesignal-sdk==1.0.0